nixos/borgbackup: replace activation script with oneshot unit
The task of ensuring various directories exist and are owned by the
appropriate user does not need to happen until the backup job is about
to start, so defer it until then rather than running sequentially in an
activation script before systemd starts.

This also prevents a backup job from starting if the user's home
directory doesn't exist, which I'd argue is better than having it fail
later.

Tested with:
  nix-build nixos/release.nix -A tests.borgbackup.x86_64-linux