nixos/ups: replace activation script with tmpfiles.d
This activation script only exists to create a directory, which is not
used until the systemd service units actually start. This is exactly the
kind of thing which tmpfiles.d exists to handle, so use that instead.

Tested by instantiating a minimal configuration.nix with:

    boot.isContainer = true;
    power.ups.enable = true;

and running the resulting container. Obviously the services didn't start
since I didn't configure any devices or provide the config files which
are supposed to contain secrets, but at least the configuration built
and booted okay.