nixos/systemd: allow combining job scripts with Exec*
This fixes a "value is a set while a list was expected" error if you try
to use both preStart and serviceConfig.ExecStartPre, for example, or
similar combinations of any of the Exec* options.

You do have to ensure that your Exec* definitions are (possibly
singleton) lists, not strings, and you have to be aware that the
relative ordering of the scripts is difficult to predict. With those
caveats, this can be a useful tool for modifying the behavior of a
service which is defined elsewhere.