nixos/networking: resolvconf doesn't need /var
As far as I can tell, this was a spurious dependency from the
"resolvconf" snippet on the "var" snippet. Amongst things in /var,
openresolv's resolvconf only uses /var/run, and only if it isn't built
to use /run directly, which it is in nixpkgs. So I think this is fine.