nixos/activation: mount ro tmpfs over /var/empty
Previously, the "var" activation script recursively deleted the contents
of /var/empty and then marked it "immutable" to make it unlikely that
any files would appear in it.

This commit instead mounts a read-only tmpfs over /var/empty, which
makes it equally unlikely that any files will appear in it. This is
something systemd knows how to do declaratively.

In either case, root could always take steps to make the directory not
be empty, but they'd have to be actually trying.