nixos/systemhealth: remove
This module appears to be unmaintained.

The upstream software was originally written for 2.6. There was a kludge
in this module to convince it to run on newer kernels:

    # Make it work for kernels 3.x, not so different than 2.6
    sed -i 's/2\.6/4.0/' system_health.py

(I love how the comment says 3.x and then somebody updated only the
implementation to 4.0.) I tried the obvious change of making that say
"5.0" instead, but on startup it still said "Sorry, kernel v5.1 is not
supported".

I'm skeptical that anybody is using this module, but if somebody is and
wants to maintain it, then they can revert this commit and fix things.